функции рендеринга в отдельном файле
